edit CartSlice: debugging AddToCart
